python/pyproject.toml
[tool.poetry]
name = "bricklens"
version = "0.1.0"
description = ""
authors = ["Matt Welsh <mdw@mdw.la>"]

[tool.poetry.dependencies]
python = ">=3.7,<3.9"
numpy = "^1.19.2"
pandas = "^1.1.5"
Pillow = "^7.2.0"
progress = "^1.5"
# pyldraw = { git = "https://github.com/mdwelsh/python-ldraw", branch = "master" }
pyldraw = { path = "../../python-ldraw/", develop = false }
pytorch-ignite = "^0.4.2"
tensorboard = "^2.4.0"
torch = "^1.7.1"
torchvision = "^0.8.2"
wandb = "^0.10.12"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^3.4"
